Refactor arch.py. Add build target name.
